Automated alembic downgrade upgrade on all databases

+++ lib/rucio/db/sqla/migrate_repo/versions/469d262be19_removing_created_at_index.py
@@ -1,39 +1,47 @@
-'''
- Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Martin Barisits <martin.barisits@cern.ch>, 2014
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
- Licensed under the Apache License, Version 2.0 (the "License");
- You may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+''' removing created_at index '''
 
- Authors:
- - Martin Barisits, <martin.barisits@cern.ch>, 2014
- - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
-
-Removing created_at index
-
-Revision ID: 469d262be19
-Revises: 16a0aca82e12
-Create Date: 2014-04-16 14:52:30.562161
-'''
+from alembic import context
 from alembic.op import create_index, drop_index
 
 
-# revision identifiers, used by Alembic.
-revision = '469d262be19'  # pylint: disable=invalid-name
-down_revision = '16a0aca82e12'  # pylint: disable=invalid-name
+revision = '469d262be19'
+down_revision = '16a0aca82e12'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    create_index('UPDATED_DIDS_SCOPERULENAME_IDX', 'updated_dids', ['scope', 'rule_evaluation_action', 'name'])
-    drop_index('CREATED_AT_IDX', 'updated_dids')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        create_index('UPDATED_DIDS_SCOPERULENAME_IDX', 'updated_dids', ['scope', 'rule_evaluation_action', 'name'])
+        drop_index('CREATED_AT_IDX', 'updated_dids')
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_index('UPDATED_DIDS_SCOPERULENAME_IDX', 'updated_dids')
-    create_index('CREATED_AT_IDX', 'updated_dids', ['created_at'])
+
+    if context.get_context().dialect.name in ['oracle', 'mysql', 'postgresql']:
+        drop_index('UPDATED_DIDS_SCOPERULENAME_IDX', 'updated_dids')
+        create_index('CREATED_AT_IDX', 'updated_dids', ['created_at'])
